# Vector Quantiser Index Optimisation