… misc/CMakeLists.txt
… octave/trellis.m
… octave/trellis_dec3_nstage3.txt
… script/subsetvq.sh
… script/train_trellis.sh