Experiments in vector index optimisation to improve robustness to bit errors. An alternative to the Trellis (Maximum Likelihood) methods explored in https://github.com/drowe67/codec2/pull/205.  Similar techniques have also been used to develop the [450 bit/s Codec 2](https://www.rowetel.com/?p=6212) mode.

# Binary Switching Example

An example of the Psuedo-Gray binary switching algorithm on a 128 entry, 2 dimensional VQ trained on Gaussian samples.  The red lines are the hamming distance 1 neighbors for the first vector:

![Screenshot from 2021-09-03 13-57-47](https://user-images.githubusercontent.com/45574645/131950552-5fe56652-6162-4b75-b531-9c2c09968598.png)

They are fairly close, but not optimal.  I guess it's hard to optimise all vectors to be the minimum distance.

# Binary Switching Usage

Binary switching with probability info.  First run training info through VQ with `--vec_usage` option.
```
cat ve9qrp_lim.f32 | ./misc/vq_mbest --st 2 --en 16 -k 20 -q vq_stage1.f32 --vec_usage > /dev/null 2>ve9qrp_vu.txt
```
Edit the `ve9qrp_vu.txt` file to remove the top 2 lines so it only contains usage info.  Load it into Octave, normalise the values (this may be optional), save to a .f32 file:
```
load ../build_linux/train_vu.txt
x=train_vu/sum(train_vu)
save_f32("../build_linux/vq_stage1_prob.f32",x)
```
Then run binary switching with `--prob` option:
```
./misc/vq_binary_switch -d 20 --st 2 --en 16 vq_stage1.f32 vq_stage1_bs004.f32 -f --prob vq_stage1_prob.f32
```
This can be allowed to run for some time (hours), however I'm not sure much the long run times improve the initial results which show significant improvement after about 10 minutes.  After a few thousand iterations a MSE ratio of < 0.30 is typical (ie the average MSE has been reduced to 0.3 of the original value for all Hamming 1 errors over the VQ).

# Binary Switching Results

Here are some AWGN channel results for the 12 bit (4096 entry by 20 dim) VQ [1][2] optimised using the binary switching algorithm.  The first plot is just the BER and PER for BPSK, provided as a way to read off BER and PER for each test point. The second plot is the RMS spectral distortion (SD) for several different cases:

no errors - best case SD with no bit errors
vanilla AWGN - original VQ before optimisation
binary switch - VQ indexes optimised using BS algorithm (with probabilities)
ldpc (112,56) - using a (112,56) LDPC code with original VQ
binary switch with ldpc (112,56) - LDPC code with binary switch optimised VQ
binary switch with trellis dec3 - using the trellis decoder method from https://github.com/drowe67/codec2/pull/205

![Screenshot from 2021-09-04 13-02-28](https://user-images.githubusercontent.com/45574645/132081078-8ade3ee8-fe82-4fc6-afbe-a67a9bcd4320.png)
![Screenshot from 2021-09-13 07-34-11](https://user-images.githubusercontent.com/45574645/133004578-32857815-1188-4e79-8f0c-1d6d2a40429b.png)

Notes:
1. The BS algorithm greatly improves SD compared to AWGN with no index optimisation, and gives results comparable to the LDPC FEC code.  However unlike FEC, it has a "long tail" - there are still some errors at higher Eb/No.
1. The Trellis results are interesting too, and deserve further investigation.
1. Using the index optimised VQ and LDPC together is very effective.
1. At low Eb/No (2dB-ish) the PER is 50%! I'm quite amazed we get any speech through when every 2nd vector has one or more bit errors.
1. This technique gives last ditch/low quality speech at quite low Eb/No.  So it pushes the lower limit down in a gradual sense, compared to the previous LDPC results that "fell over" at about Eb/No 2dB as errors crept into the decoded speech.
1. At low Eb/No the speech quality using binary switching is degraded, but still usable.  Unlike FEC it doesn't correct errors, but makes them less noticeable.
1. Lowering the minimal usable Eb/No is quite useful for fading channels, better to get some speech through at the bottom of fades than lose it entirely.  This is very analog, "brain FEC" will help put it all back together.

(112,56) LDPC code performance for comparison to SD against Eb/No above.  Note the location of the "knee" between 2 and 3dB.

![Screenshot from 2021-09-11 10-13-43](https://user-images.githubusercontent.com/45574645/132930652-8a54983e-97f4-48a2-bbd5-b87bc7010bf6.png)

# Samples

 | Codec Mode | Eb/No dB | BER | VQ | Decoder | Sample |
| --- | --- | --- | --- | --- | --- |
| 600 | high | 0.0 | Original | Normal | [listen](http://rowetel.com/downloads/codec2/vq_sort/big_dog_10dB.wav) |
| 600 | 1 | 0.06 | Original | Normal | [listen](http://rowetel.com/downloads/codec2/vq_sort/big_dog_1dB.wav) |
| 600 | 1 | 0.06 | binary switch | Normal | [listen](http://rowetel.com/downloads/codec2/vq_sort/big_dog_binsw_1dB.wav) |
| 600 | 1 | 0.06 | binary switch | Trellis| [listen](http://rowetel.com/downloads/codec2/vq_sort/big_dog_trellis_1dB.wav) |
| 700C | high | 0.0 | Original | Normal | [listen](http://rowetel.com/downloads/codec2/vq_sort/big_dog_700c_10dB.wav) |
| 700C | 1 | 0.06 | Original | Normal | [listen](http://rowetel.com/downloads/codec2/vq_sort/big_dog_700c_1dB.wav) |
| 450 | high | 0.0 | Original | Normal | [listen](http://rowetel.com/downloads/codec2/vq_sort/big_dog_450_10dB.wav) |
| 450 | 1 | 0.06 | Original | Normal | [listen](http://rowetel.com/downloads/codec2/vq_sort/big_dog_450_1dB.wav) |

Notes:
1. Just a single sample to get a feel for where this work is going, more samples would help us get a clearer picture.
1. "600" is a tag for the sort of Codec 2 mode we could build with the 12 bit quantiser (see Applications below).
2.  The level isn't quite right so "600" isn't as loud as 700C and 450.
3. I inserted bit errors in the 700C and 450 samples using the `c2dec --ber` option.
4. The "ch" at the start of 450 samples might be the wave file header.

# Synthesising Samples

1. For these tests I've settled on 30ms (dec=3) frame update rate (12 bits/fr/0.03s=400 bits/s for the VQ info).  Higher update rates are more robust (less sensitive to bit errors at the same Eb/No or BER), and also sounds better without errors.
1. The 10% PER and 1% BER rule still applies for "readability 5" speech, but at low Eb/Nos (e.g. 1dB) the speech is still "readability 3" - readable with some difficulty.

Script to synthesise samples:
```
cd codec2/build_linux
../script/subsetvq.sh -x -v vq_stage1_bs004.f32 ../wav/ve9qrp.wav aplay -d 3 -e 1
<snip>
Eb/No: 1.00 dB nframes: 3749 nerrors: 2523 BER: 0.056 PER: 0.50 mse: 13.69 79.98
```
This examples uses the `vq_stage1_bs004.f32` VQ that has been optimised with the binary switch algorithm using the probabilities, decimation 3, Eb/No of 1dB.  The output is played directly through the sound card.  I estimate the quality to be "readability 3 - readable with difficulty".  So you can get info through using phonetics and repeats.

# Applications

1. Index optimisation could be tried on Codec 2 700C, it could lead to improved performance at low SNRs for FreeDV 700C/D/E simply by swapping in new VQ tables.  This would break compatibility with the current modes.  Not hard to try, as it's just requires index optimisation of the existing VQ.  It could also be applied to FreeDV 2020, which uses a 4 stage VQ.

1. The 12 bit VQ used for these tests could be used for a new Codec 2 mode.  Assuming 6 bits for pitch and a 30ms update rate that's (12+6)/0.03 = 600 bits/s.  Using the binary switch index optimisation (with or without LDPC), and assuming 2dB for modem overheads we could get "readability 3" speech through at Eb/No=1dB, so a SNR = 1 + 10*log10(600/3000) + 2 = -4dB.  FreeDV 700D currently has a threshold of around -2dB.  However it would be worth comparing over the air and on different fading channels to get a full comparison, e.g. in error free speech quality (30ms versus 40ms update rate), and performance coming in and out of fades.

# References

[1] https://github.com/drowe67/codec2/pull/151
[2] https://www.rowetel.com/?p=7754